hv: vm_config: enable sgx support for uos in sdc scenario
This is a example to enable sgx support for a guest.
Note sgx virtualization is not supported in SOS VM, please don't add epc
config in SOS VM.
For UOS, the epc base should be 0x80000000, which is aligned with DM to reserve the
memory range in E820 table; the epc size should be not greater than the physcial EPC
size.

Signed-off-by: Binbin Wu <binbin.wu@intel.com>