hv: assign: use the irte index directly if valid
For MSI, hypervisor will reserve 2^n (n from 0 to 5) continuous irte for the
device use the MSI.
Add a new paramter to ptirq_prepare_msix_remap to specifiy if the ptirq should
use the pre-allocated irtes.
When the paramter irte_idx >=0, it is valid, use it directly for the ptirq.
When the paramter irte_idx < 0, it is not valid.

- MSI: will alway use the reserved IRTEs (refer to the next patch)
  These reserved IRTE is bound to the ptdev, will not be released after setup.

- MSI-X / GSI:
  - allocate a new IRTE for the ptirq entry when now IRTE is associated
    with the ptirq entry.
  - reuse the IRTE if the ptirq entry already has a IRTE.
  - release the IRTE when the ptirq entry is released.

Signed-off-by: Binbin Wu <binbin.wu@intel.com>